requirements.txt
tensorflow~=2.8.0
numpy~=1.22.2
matplotlib~=3.5.1
Pillow~=9.0.1
sklearn~=0.0
scikit-learn~=1.0.2
pycoral~=2.0.0
keras~=2.8.0